made quick and dirty module for saving the resampling errors to file. currently only prints the std. error found. I should be able to add eventual other values, but I'm unsure of how or which.
the bootstrap code finds both the bias and the std. error. It also
computes the mean of the sampled values. I don't know if this is worth
keeping and I should look into eventual bias wrt. the blocking method,
or if the blocking method is intended to work on so much data that the
bias would mean holding a too large array in memory or something
similar. the error estimates from blocking are now at least in the
results folder.